Added ThreadedCdepIncludeDependencyGenerator doctests and removed duplication with DirwiseCdepIncludeDependencyGenerator
